requirements.txt
fastapi
uvicorn
pillow
transformers
diffusers
torch
sentence-transformers
